requirements.txt
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Django==2.0.5
django-widget-tweaks==1.4.2
# django-wordpress-auth==0.1.31
EasyProcess==0.2.3
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
# phpserialize==1.3
pycountry==18.5.26
pylint==1.9.2
python-dateutil==2.7.3
pytz==2018.4
PyVirtualDisplay==0.2.1
requests==2.18.4
selenium==3.12.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11

pinax-stripe
pinax-templates
